+++ __mocks__/ol/Observable.ts
@@ -1,0 +1,3 @@
+export const unByKey = jest.fn();
+
+export default {};